# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Distribution Management System API
  version: 1.0.0
  description: |
    Partial OpenAPI definition for the Distribution Management System.
    This file includes core auth and order/customer/rep endpoints and representative schemas.
servers:
  - url: http://localhost:5000
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          description: payload
...
paths:
...
  /api/admin/reps:
    get:
      summary: Get all reps (Admin)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Paged reps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
security:
  - BearerAuth: []
...